Module Functoria_runtime

Functoria runtime.

module Arg : sig ... end

Arg defines command-line arguments which can be set at runtime. This module is the runtime companion of Functoria_key. It exposes a subset of Cmdliner.Arg.

module Key : sig ... end

Key defines values that can be set by runtime command-line arguments. This module is the runtime companion of Functoria_key.

val with_argv : unit Cmdliner.Term.t list -> string -> string array -> unit

with_argv keys name argv evaluates the keys terms on the command-line argv. name is the executable name. On evaluation error the application calls exit(3) with status 64. If `Help or `Version were evaluated, exit(3) is called with status 63.